requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
apache-airflow==2.8.1
apache-airflow-providers-cncf-kubernetes==7.4.2
kubernetes==28.1.0
